+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.3
+**features**
+- Add support for 'count' property (display a counter with number of typed characters) 
+
+**technical improvements**
+- Refactor rule conversion
+
 ## 1.0.2
 **technical improvements**
 - Use new github repository for reflux-component